+++ RELEASE_NOTES.md
@@ -1,5 +1,6 @@
 #### 5.0.0-beta.16
 * Make Adapter.getLogger a verbose log operation
+* Fix #347 - BufferFull result from Logary proper, through Facade cast exn
 
 #### 5.0.0-beta.15
 * Fix #341 — template from context as well as fields